+++ packages/types/src/sdk/index.ts
@@ -25,3 +25,4 @@ export * from "./permissions"
 export * from "./row"
 export * from "./vm"
 export * from "./view"
+export * from "./oauth2"
